+++ src/services/ClearCacheService.php
@@ -57,8 +57,6 @@ class ClearCacheService extends Component
 
         Blitz::$plugin->cachePurger->purgeUris($event->siteUris);
 
-        Blitz::$plugin->deployer->deployUris($event->siteUris);
-
         if ($this->hasEventHandlers(self::EVENT_AFTER_CLEAR_CACHE)) {
             $this->trigger(self::EVENT_AFTER_CLEAR_CACHE, $event);
         }
@@ -91,8 +89,6 @@ class ClearCacheService extends Component
         Blitz::$plugin->cacheStorage->deleteAll();
 
         Blitz::$plugin->cachePurger->purgeAll();
-
-        Blitz::$plugin->deployer->deployAll();
 
         if ($this->hasEventHandlers(self::EVENT_AFTER_CLEAR_ALL_CACHE)) {
             $this->trigger(self::EVENT_AFTER_CLEAR_ALL_CACHE, $event);